Finished ProductController

+++ rest/src/main/java/cz/ucl/controller/ProductController.java
@@ -3,10 +3,12 @@ package cz.ucl.controller;
 import cz.ucl.model.product.Product;
 import cz.ucl.service.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpEntity;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 public class ProductController {
@@ -20,9 +22,16 @@ public class ProductController {
     }
 
     @GetMapping(value = "/products/{id}")
-    public String getProduct(@PathVariable Integer id) {
-        return productService.getProductFromId(id).toString();
+    @ResponseBody
+    public ResponseEntity<Object> getProduct(HttpEntity<String> httpEntity, @PathVariable Integer id) {
+        if (productService.getProductFromId(id).isEmpty()){
+            return ResponseEntity.status(400).body("No product with this ID.");
+        }
+        else {
+            return ResponseEntity.status(200).body(productService.getProductFromId(id));
+        }
     }
+
 
     @PostMapping("/products")
     public ResponseEntity<Object> addProduct(@RequestBody Product product){
@@ -42,11 +51,46 @@ public class ProductController {
         return ResponseEntity.status(200).body(product);
     }
 
+    @PostMapping("/products/{id}")
+    public ResponseEntity<Object> updateProduct(@RequestBody Product product, @PathVariable int id){
+        //Validations
+        if (productService.getProductFromId(id).isEmpty()){
+            return ResponseEntity.status(400).body("Product with this ID does not exist");
+        }
+        if (product.getPrice() > 100000.00 || product.getPrice() < 0){
+            return ResponseEntity.status(400).body("Price is below 0 or higher than 100000");
+        }
+        if (product.getAvailable() < 0){
+            return ResponseEntity.status(400).body("There must be at least one product");
+        }
+
+        Optional<Product> optionalProduct = productService.getProductFromId(id);
+        Product newProduct = optionalProduct.get();
+
+        newProduct.setPrice(product.getPrice());
+
+        if (product.getName() != null && !product.getName().equals("")) {
+            newProduct.setName(product.getName());
+        }
+        if (product.getDescription() != null && !product.getDescription().equals("")){
+            newProduct.setDescription(product.getDescription());
+        }
+        newProduct.setAvailable(product.getAvailable());
+
+        Product toShow = productService.addProduct(product);
+        return ResponseEntity.status(200).body(toShow);
+    }
 
 
     @DeleteMapping(value = "/products/{id}")
-    public void deleteProduct(@PathVariable Integer id){
-        productService.deleteProduct(id);
+    public ResponseEntity<Object> deleteProduct(HttpEntity<String> httpEntity, @PathVariable Integer id){
+        if (productService.getProductFromId(id).isEmpty()){
+            return ResponseEntity.status(400).body("Product with this ID does not exist");
+        }
+        else {
+            productService.deleteProduct(id);
+            return ResponseEntity.status(200).build();
+        }
     }
 
 }
